+++ html_form_builder/doc/changelog.rst
@@ -1,2 +1,11 @@
-v1.1 Added Submit action feature
-v1.0 Initial Release
+v1.3
+====
+* Add 2 fields types, dropbox(many2one) and radio button group(selection)
+
+v1.1
+====
+* Added Submit action feature
+
+v1.0
+====
+* Initial Release